Trim down the list of recommended streams (#235)
* Trim down the list of recommended streams

Besides including pinned streams, this will trim down the list of recommended streams up to the latest stable version.

* Improved snippet

The selected code is checking if there are any stable versions in the platformReleases list. If there are none, it adds the current item to the list. This code can be improved by using the anyMatch method instead of filter and count. The anyMatch method will return as soon as it finds a match, which can be more efficient than counting all matches with filter and count.